requirements.txt
aniso8601==8.0.0
cbor2==5.4.2.post1
certifi==2019.9.11
chardet==3.0.4
click==8.1.1
colorama==0.4.4
colorzero==1.1
Deprecated==1.2.13
Flask==2.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.7
Flask-SocketIO==4.2.1
gpiozero==1.5.1
idna==2.8
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
luma.core==2.3.1
luma.oled==3.8.1
MarkupSafe==2.1.1
pigpio==1.44
Pillow==9.1.0
pkg_resources==0.0.0
pyftdi==0.54.0
pyserial==3.5
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.2
pyusb==1.2.1
requests==2.22.0
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.4.1
spidev==3.5
urllib3==1.25.3
waitress==2.1.1
Werkzeug==2.1.0
wrapt==1.14.0
zipp==3.8.0
